[CU-2mtr0gw] Update submission (#13)
* Update submission functionality and fixed authorization of actions

* Fixed test

* 403 messages

+++ app/Http/Controllers/DeleteSubmissionController.php
@@ -10,7 +10,9 @@ class DeleteSubmissionController
 {
     public function __invoke(Submission $submission): JsonResponse
     {
-        Auth::user()->can('delete', $submission);
+        if (Auth::user()->cannot('delete', $submission)) {
+            abort(403, 'You dont own this submission');
+        }
         $submission->delete();
 
         return response()->json([
